Move all filters to their own package
