+++ pkg/ctl/cmdutils/configfile_test.go
@@ -9,6 +9,7 @@ import (
 
 	api "github.com/weaveworks/eksctl/pkg/apis/eksctl.io/v1alpha5"
 	. "github.com/weaveworks/eksctl/pkg/ctl/cmdutils"
+	"github.com/weaveworks/eksctl/pkg/ctl/cmdutils/filter"
 )
 
 var _ = Describe("cmdutils configfile", func() {
@@ -125,7 +126,7 @@ var _ = Describe("cmdutils configfile", func() {
 				}
 
 				params := &CreateClusterCmdParams{WithoutNodeGroup: true, Managed: false}
-				Expect(NewCreateClusterLoader(cmd, NewNodeGroupFilter(), nil, params).Load()).To(Succeed())
+				Expect(NewCreateClusterLoader(cmd, filter.NewNodeGroupFilter(), nil, params).Load()).To(Succeed())
 				cfg := cmd.ClusterConfig
 				Expect(cfg.VPC.NAT.Gateway).To(Not(BeNil()))
 				Expect(*cfg.VPC.NAT.Gateway).To(Equal(natTest.expectedGateway))
@@ -174,7 +175,7 @@ var _ = Describe("cmdutils configfile", func() {
 					ProviderConfig: &api.ProviderConfig{},
 				}
 
-				ngFilter := NewNodeGroupFilter()
+				ngFilter := filter.NewNodeGroupFilter()
 
 				Expect(cmd.ClusterConfig.NodeGroups).To(HaveLen(0))
 
@@ -238,7 +239,7 @@ var _ = Describe("cmdutils configfile", func() {
 					ProviderConfig:    &api.ProviderConfig{},
 				}
 
-				ngFilter := NewNodeGroupFilter()
+				ngFilter := filter.NewNodeGroupFilter()
 
 				params := &CreateClusterCmdParams{
 					WithoutNodeGroup: loaderTest.withoutNodeGroup,
@@ -277,7 +278,7 @@ var _ = Describe("cmdutils configfile", func() {
 					Managed:          false,
 				}
 
-				Expect(NewCreateClusterLoader(cmd, NewNodeGroupFilter(), nil, params).Load()).To(Succeed())
+				Expect(NewCreateClusterLoader(cmd, filter.NewNodeGroupFilter(), nil, params).Load()).To(Succeed())
 				cfg := cmd.ClusterConfig
 				assertValidClusterEndpoint(cfg.VPC.ClusterEndpoints, expectedPrivAccess, expectedPubAccess)
 			}
