+++ pkg/ctl/cmdutils/filter/filter_suite_test.go
@@ -1,0 +1,11 @@
+package filter
+
+import (
+	"testing"
+
+	"github.com/weaveworks/eksctl/pkg/testutils"
+)
+
+func TestSuite(t *testing.T) {
+	testutils.RegisterAndRun(t)
+}
