+++ pkg/ctl/cmdutils/filter/nodegroup_filter_test.go
@@ -1,4 +1,4 @@
-package cmdutils
+package filter
 
 import (
 	"bytes"
